rename directory to be consistent with other names
